{{cookiecutter.project_name}}/requirements.txt
celery==5.4.0
django==5.0.6
django-cors-headers==4.4.0
django-extensions==3.2.3
djangorestframework==3.15.2
gunicorn==22.0.0
psycopg==3.1.19
